+++ assets/src/components/activities/check_all_that_apply/delivery-entry.ts
@@ -1,5 +1,5 @@
 // This is the entry point for the delivery component for
-// the multiple choice activitity.  All that needs to be exposed
+// the check all that apply activitity.  All that needs to be exposed
 // is the actual delivery web component.
 
 export { CheckAllThatApplyDelivery } from './CheckAllThatApplyDelivery';