+++ assets/src/components/editing/editor/Editor.tsx
@@ -19,6 +19,7 @@ import { withVoids } from './overrides/voids';
 import { withInlines } from './overrides/inlines';
 import { withTables } from './overrides/tables';
 import { withMarkdown } from './overrides/markdown';
+import { onPaste } from './handlers/paste';
 
 export type EditorProps = {
   // Callback when there has been any change to the editor (including selection state)
@@ -44,6 +45,8 @@ function areEqual(prevProps: EditorProps, nextProps: EditorProps) {
 }
 
 export const Editor = React.memo((props: EditorProps) => {
+
+  const [isPerformingAsyncAction, setIsPerformingAsyncAction] = useState(false);
 
   const commandContext = props.commandContext;
 
@@ -112,8 +115,10 @@ export const Editor = React.memo((props: EditorProps) => {
         onChange={onChange}>
 
         <InsertionToolbar
+          isPerformingAsyncAction={isPerformingAsyncAction}
           toolbarItems={props.toolbarItems}
-          commandContext={props.commandContext} />
+          commandContext={props.commandContext}
+        />
 
         <HoveringToolbar
           isOpen={shouldShowFormattingToolbar}>
@@ -128,7 +133,12 @@ export const Editor = React.memo((props: EditorProps) => {
           renderElement={renderElement}
           renderLeaf={renderLeaf}
           placeholder="Enter some content here..."
-          onKeyDown={onKeyDown} />
+          onKeyDown={onKeyDown}
+          onPaste={async (e) => {
+            setIsPerformingAsyncAction(true);
+            await onPaste(editor, e, props.commandContext.projectSlug);
+            setIsPerformingAsyncAction(false);
+          }} />
       </Slate>
     </React.Fragment>
   );