+++ lib/oli/lti_1p3/registration.ex
@@ -20,7 +20,25 @@ defmodule Oli.Lti_1p3.Registration do
   @doc false
   def changeset(registration, attrs \\ %{}) do
     registration
-    |> cast(attrs, [:issuer, :client_id, :key_set_url, :auth_token_url, :auth_login_url, :auth_server, :tool_jwk_id, :institution_id])
-    |> validate_required([:issuer, :client_id, :key_set_url, :auth_token_url, :auth_login_url, :auth_server, :tool_jwk_id, :institution_id])
+    |> cast(attrs, [
+      :issuer,
+      :client_id,
+      :key_set_url,
+      :auth_token_url,
+      :auth_login_url,
+      :auth_server,
+      :tool_jwk_id,
+      :institution_id,
+    ])
+    |> validate_required([
+      :issuer,
+      :client_id,
+      :key_set_url,
+      :auth_token_url,
+      :auth_login_url,
+      :auth_server,
+      :tool_jwk_id,
+      :institution_id
+    ])
   end
 end