+++ lib/oli_web/views/layout_view.ex
@@ -27,8 +27,18 @@ defmodule OliWeb.LayoutView do
     if active == path do :active else nil end
   end
 
-  def sidebar_link(%{:assigns => assigns} = _conn, text, path, [to: route]) do
-    link text, to: route, class: active_class(active_or_nil(assigns), path)
+  def sidebar_link(%{:assigns => assigns} = _conn, text, path, opts) do
+    route = Keyword.get(opts, :to)
+    badge = Keyword.get(opts, :badge)
+
+    case badge do
+      nil ->
+        link text, to: route, class: active_class(active_or_nil(assigns), path)
+      badge ->
+        link to: route, class: "align-items-center #{active_class(active_or_nil(assigns), path)}" do
+          [content_tag(:span, text), content_tag(:span, badge, class: "badge badge-pill badge-primary ml-2")]
+        end
+    end
   end
 
   def is_admin?(%{:assigns => assigns}) do