+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
+## Unreleased
+### Enhancements
+  - Improved LTI workflow for new institutions
+  - Add support for embedding images in structured content editors by external URL and by pasting a copied image
+  - Ordering activity
+  - Add support for user help requests capture and forward to email or help desk
+
+### Bug fixes
+  - Fix a broken link to external learning objectives content
 
 ## 0.4.1 (2021-1-4)
 ### Bug fixes